Decompose the RightSide.jsx for readability.
